Implementing trap collision.
Due to the interaction between the robot and traps I've been tweaking the trap class, and there is a rough collision detection that has been implemented between them. I've also updated the trap spawn function to ensure that it doesn't load in any given landmark. (Please ignore the odd placement of landmarks as it allows me to test the 'bypassLandmark' function.